requirements.txt
Faker==0.7.11
phonenumberslite==8.4.2
